ci: Use github-actions[bot] as format commits user (#19)
Currently, the Format Pipeline will automatically format the Terraform
code and commit changes directly to the source branch for a PR. However,
this commit will not trigger another status check run.

To address this, we must add a token to the checkout action. When doing
this, we would also like to migrate away from using an upstream Action
and use the `github-actions[bot]` user directly.